Extract out parsing logic to new class
