+++ src/main/java/TaskDecoder.java
@@ -3,7 +3,6 @@ import java.time.format.DateTimeFormatter;
 
 public class TaskDecoder {
     
-
     public static Task retTask(String taskDesc) {
         String[] splits = taskDesc.split("\\|");
         String tags = splits[0];
